added /grid and /collapse endpoints

+++ app.py
@@ -1,7 +1,7 @@
 import os
 import os.path
 from distutils.dir_util import mkpath, copy_tree, remove_tree
-from flask import Flask, render_template
+from flask import Flask, render_template, url_for
 from wave_function_processing import Grid, GRID_PX_SIZE, TILE_PX_SIZE, \
                                      IMAGE_DICT, NEIGHBOURS_DICT
 
@@ -9,11 +9,12 @@ TILES_FOLDER = "tiles/"
 OUTPUT_FOLDER = "output/"
 
 # Creating cache folder for users generation
-WORKING_FOLDER = "tmp/"
+WORKING_FOLDER = "./tmp/"
 WORKING_TILES_FOLDER = os.path.join(WORKING_FOLDER, TILES_FOLDER)
 WORKING_OUTPUT_FOLDER = os.path.join(WORKING_FOLDER, OUTPUT_FOLDER)
 
-remove_tree(WORKING_FOLDER)
+if os.path.exists(WORKING_FOLDER):
+    remove_tree(WORKING_FOLDER)
 mkpath(WORKING_FOLDER)
 copy_tree(TILES_FOLDER, WORKING_TILES_FOLDER)
 copy_tree(OUTPUT_FOLDER, WORKING_OUTPUT_FOLDER)
@@ -31,9 +32,16 @@ def show_index():
     return render_template("index.html")
 
 
-@app.route('grid', methods=['GET'])
+@app.route('/grid', methods=['GET'])
 def show_grid():
     # Showing current state of grid
     filename = "cur_grid.png"
-    grid.save_png(filename, folder="WORKING_OUTPUT_FOLDER")
-    return
+    grid.save_png(filename, folder="static")
+    return render_template("grid.html", grid_image=filename)
+
+
+@app.route('/collapse', methods=['GET', 'POST'])
+def collapse():
+    # Collapse one cell on the grid and show current state
+    grid.collapse_next_cell()
+    return show_grid()
